+++ RELEASENOTES.md
@@ -3,6 +3,12 @@ Release Notes
 
 ## v1.14.0
 
+#### v1.14.1
+- updated dependencies
+- extended docstore test
+- adapted linter config
+- bstring type tests
+
 ### Changes
 
 - Added support of Go 1.25.